favorite movies

+++ src/components/profile-view/profile-view.jsx
@@ -1,18 +1,21 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { Form, Button, Col, Row, Container } from 'react-bootstrap';
+import { FavoriteMovies } from '../../favorite-movies/favorite-movies';
+
+import { Link } from 'react-router-dom';
 
 export function ProfileView(props) {
-  const [user, setUser] = useState(props.user);
-  const [favoriteMovies, setFavoriteMovies] = useState(props.favoriteMovies);
+  const [user, setUser] = useState('');
+  const [favoriteMovies, setfavoriteMovies] = useState('');
   const [email, setEmail] = useState(props.email);
-  const [birth_date, setBirth_date] = useState(props.birth_date);
+  const [birthday, setBirth_date] = useState(props.birthday);
 
   // hooks for user inputs
   const [values, setValues] = useState({
     usernameErr: '',
     passwordErr: '',
-    birth_dateErr: '',
+    birthdayErr: '',
     emailErr: '',
   });
 
@@ -26,7 +29,7 @@ export function ProfileView(props) {
       })
       .then((response) => {
         setUser(response.data);
-        setFavoriteMovies(response.data.FavoriteMovies);
+        setfavoriteMovies(response.data.FavoriteMovies);
       })
       .catch((error) => console.error('getUser Error ' + error));
   };
@@ -36,16 +39,21 @@ export function ProfileView(props) {
   }, []);
 
   const handleDelete = () => {
-    axios
-      .delete(`https://my-movie-db22.herokuapp.com/users/${user._id}`, {
-        headers: { Authorization: `Bearer ${token}` },
-      })
-      .then(() => {
-        alert(`The account ${user.Username} was successfully deleted.`);
-        localStorage.clear();
-        window.open('/register', '_self');
-      })
-      .catch((error) => console.error('handleDelete Error ' + error));
+    if (user && token) {
+      let sure = confirm('Are you sure? This action is irreversible and will ERASE your account.');
+      if (!sure) return;
+      // request to Delete user from webserver
+      axios
+        .delete(`https://my-movie-db22.herokuapp.com/users/${user._id}`, {
+          headers: { Authorization: `Bearer ${token}` },
+        })
+        .then(() => {
+          alert(`The account ${user.Username} was successfully deleted.`);
+          localStorage.clear();
+          window.open('/register', '_self');
+        })
+        .catch((error) => console.error('handleDelete Error ' + error));
+    }
   };
 
   // user validation
@@ -86,7 +94,7 @@ export function ProfileView(props) {
             Username: username,
             Password: password,
             Email: email,
-            Birth_date: birth_date,
+            Birthday: birthday,
           },
           {
             headers: { Authorization: `Bearer ${token}` },
@@ -103,41 +111,61 @@ export function ProfileView(props) {
         });
     }
   };
-
+  if (!user) 'loading..';
   console.log(user.Email);
-  console.log(props);
+  console.log(props.movies);
+  console.log(favoriteMovies);
   return (
     <Container className="profile-view">
-      <Form.Group className="mt-3" as={Row} controlId="formUsername">
-        <Form.Label column="true">Username:</Form.Label>
-        <Col>
-          <Form.Control variant="success" type="text" value={currentUser}></Form.Control>
-        </Col>
-      </Form.Group>
-      <Form.Group className="mt-3" as={Row} controlId="formUsername">
-        <Form.Label column="true">Password:</Form.Label>
-        <Col>
-          <Form.Control variant="success" type="text" value="******"></Form.Control>
-        </Col>
-      </Form.Group>
-      <Form.Group className="mt-3" as={Row} controlId="formUsername">
-        <Form.Label column="true">E-Mail:</Form.Label>
-        <Col>
-          <Form.Control variant="success" type="email" value={user.Email}></Form.Control>
-        </Col>
-      </Form.Group>
-      <Form.Group className="mt-3" as={Row} controlId="formUsername">
-        <Form.Label column="true">Birthday:</Form.Label>
-        <Col>
-          <Form.Control variant="success" type="date" value={user.Birth_date}></Form.Control>
-        </Col>
-      </Form.Group>
-      <Button type="button" onClick={handleDelete}>
-        Delete
+      <h4>
+        Profile of <strong>{currentUser}</strong>
+      </h4>
+      <Form column="true">
+        <Form.Group className="mt-3" as={Row} controlId="formUsername">
+          <Form.Label column="true">Username:</Form.Label>
+          <Col>
+            <Form.Control
+              variant="success"
+              type="text"
+              defaultvalue={currentUser}
+              onChange={(e) => setUser}></Form.Control>
+          </Col>
+        </Form.Group>
+        <Form.Group className="mt-3" as={Row} controlId="formUsername">
+          <Form.Label column="true">Password:</Form.Label>
+          <Col>
+            <Form.Control variant="success" type="text" value="******"></Form.Control>
+          </Col>
+        </Form.Group>
+        <Form.Group className="mt-3" as={Row} controlId="formUsername">
+          <Form.Label column="true">E-Mail:</Form.Label>
+          <Col>
+            <Form.Control variant="success" type="email" value={user.Email}></Form.Control>
+          </Col>
+        </Form.Group>
+        <Form.Group className="mt-3" as={Row} controlId="formUsername">
+          <Form.Label column="true">Birthday:</Form.Label>
+          <Col>
+            <Form.Control
+              variant="success"
+              type="date"
+              value={user.Birth_date}
+              onChange={(e) => updateUser}></Form.Control>
+          </Col>
+        </Form.Group>
+      </Form>
+      <Button className="mb-3" type="button" onClick={handleDelete}>
+        <strong> Delete </strong> my profile
       </Button>
-      <Button type="button" onClick={updateUser}>
-        Update
+      <Button className="mb-3" type="button" onClick={updateUser}>
+        <strong>Update </strong> my profile
       </Button>
+      <h4>My favorite movies:</h4>
+      <Row className="justify-content-center mt-3">
+        {/* {favoriteMovies.map((m) => {
+          return <FavoriteMovies movieData={props.movies} />;
+        })} */}
+      </Row>
     </Container>
   );
 }
